Adding appointment model
Updating folder structure
