--- server/model/users.py
@@ -1,21 +1,0 @@
-from server.config import db
-
-class Users(db.Model):
-    __tablename__ = 'users'
-
-    id = db.Column(db.Integer, primary_key=True)
-    name = db.Column(db.String(64))
-    email = db.Column(db.String(64), unique=True, index=True)
-    username = db.Column(db.String(64), unique=True, index=True)
-    password_hash = db.Column(db.String(128))
-    eventType = db.relationship("EventTypes", backref="user")
-
-    def __init__(self, name, email, username, password_hash):
-        self.name = name
-        self.email = email
-        self.username = username
-        self.password_hash = password_hash
-
-    def __repr__(self):
-        return f"User - id:{self.id}, name:{self.name}, " \
-               f"email:{self.email}, username:{self.username}"
